Added collapse functionality
I spent a good portion of the day adding a custom javascript collapse 
for the catalog that worked real stupid and was kinda wonky. But then I 
found a javascript package that has this kind of thing built in to it. 

Now when you click on a picture in the catalog, it should drop down the 
the information for that bread. I'm going to try and add some more 
javascript and php functionality to the page that will add to cart and 
such tonight. After that I will upload the folder tomorrow night to the 
web service and all should be goooood.