Order functionality implemented!
I added buttons for each bread to add or subtract an amount of bread. 
That fills in the panel on the right, which when submitted, will send 
the values to the correct places in the order page.

I had to make a couple changes to the order page for it to work 
correctly. I changed the 'select' inputs to 'number' inputs, and I added 
some input validation to check if the values were initialized through 
post (that way if you just click on the order button in nav everything 
gets initialized to 0).

As it stands, I think we're all ready to set up the web page. I'll get 
to doing that Monday afternoon or evening. That way we should be all 
set!