Getting stuff working on the website.
Good news, everything (except the catalog page) is working on the 
website. Bad news is that the catalog page isn't working yet BUT I know 
what I need to fix in order to get it working. I'll be able to patch it 
up tonight and do a final push so that we both have the finished folder.

the website is https://breadsite.000webhostapp.com/index.php